:truck: files renamed following Google TypeScript styling conventions
all file names are renamed as snake_case and modules folder created to contain modules of the library
