+++ src/index.ts
@@ -1,6 +1,6 @@
-export { default as FB } from './lib/models/fb';
+export { default as FB } from './lib/modules/fb';
 export { default as Configuration } from './lib/models/configuration';
 export { default as Options } from './lib/models/options';
-export { default as InitialisationError } from './lib/errors/initialisationError';
-export { default as TwoFARequiredError } from './lib/errors/twoFARequiredError';
-export { default as Group } from './lib/models/group';
+export { default as InitialisationError } from './lib/errors/initialisation_error';
+export { default as TwoFARequiredError } from './lib/errors/two_fa_required_error';
+export { default as Group } from './lib/modules/group';
