--- src/lib/errors/initialisationError.ts
@@ -1,3 +1,0 @@
-export default class InitialisationError extends Error {
-
-}
