+++ src/lib/errors/initialisation_error.ts
@@ -1,0 +1,3 @@
+export default class Initialisation_error extends Error {
+
+}
