--- src/lib/errors/twoFARequiredError.ts
@@ -1,3 +1,0 @@
-export default class TwoFARequiredError extends Error {
-
-}
