+++ src/lib/errors/two_fa_required_error.ts
@@ -1,0 +1,3 @@
+export default class Two_fa_required_error extends Error {
+
+}
